Refactor and split unit tests
 - move some files out of util to ensure that files in utils do not
   include any files outside of utils in the project (this cleans up
   the dependency graph of the application)
 - split the tests in util/model/control
 - add custom target for building all unit-tests
 - narrow down dependencies for util tests such that only util files
   are built to run unit-tests-util (this improves build times
   substantially when working on util files)
 - update README
