+++ azure-pipelines/continuous-integration.yml
@@ -28,7 +28,9 @@ stages:
               sudo locale-gen
             displayName: 'Generate test locales'
           - bash: |
-              cmake --build . --target test-units
+              cmake --build . --target unit-tests-util
+              cmake --build . --target unit-tests-model
+              cmake --build . --target unit-tests-control
               CI=true ctest --verbose
             workingDirectory: ./build
             displayName: 'Run tests'
@@ -44,7 +46,9 @@ stages:
               cmake_flags: '-GNinja -DCMAKE_DEBUG_INCLUDES_LDFLAGS=ON -DENABLE_GTEST=ON -DDEBUG_COMPILE=ON'
           - script: |
               set PATH=%PATH%;C:\msys64\usr\bin;C:\msys64\mingw64\bin"
-              C:\msys64\usr\bin\bash -lc "cmake --build . --target test-units"
+              C:\msys64\usr\bin\bash -lc "cmake --build . --target unit-tests-util"
+              C:\msys64\usr\bin\bash -lc "cmake --build . --target unit-tests-model"
+              C:\msys64\usr\bin\bash -lc "cmake --build . --target unit-tests-control"
               C:\msys64\usr\bin\bash -lc "CI=true ctest --verbose"
             workingDirectory: ./build
             env:
@@ -65,7 +69,9 @@ stages:
               generation_path: '/Users/xournal-dev'
           - bash: |
               export PATH="$HOME/.local/bin:/Users/xournal-dev/gtk/inst/bin:$PATH"
-              cmake --build . --target test-units
+              cmake --build . --target unit-tests-util
+              cmake --build . --target unit-tests-model
+              cmake --build . --target unit-tests-control
               CI=true ctest --verbose
             workingDirectory: ./build
             displayName: 'Run tests'
