+++ src/control/CrashHandler.cpp
