+++ src/control/CrashHandlerWindows.h
