+++ src/control/settings/DeviceListHelper.cpp
