+++ src/control/settings/DeviceListHelper.h
