+++ src/gui/IconNameHelper.cpp
