+++ src/gui/IconNameHelper.h
